refactor(KakaoMap) : #365 - redux 분리
redux 분리, 타입 분리

Related to: #365
