--- Frontend/nubio/src/types/RootState.d.ts
@@ -1,9 +1,0 @@
-import KakaoSlice from "../redux/slice/KakaoSlice";
-import EnjoySlice from "../redux/slice/EnjoySlice";
-
-export type RootState = {
-  search: ReturnType<typeof KakaoSlice>;
-  enjoy: ReturnType<typeof EnjoySlice>;
-};
-
-export default RootState;
